+++ src/generators/Plugin.php
@@ -9,6 +9,7 @@ namespace craft\generator\generators;
 
 use Composer\Json\JsonManipulator;
 use Composer\Semver\Comparator;
+use Composer\Semver\VersionParser;
 use Craft;
 use craft\base\Model;
 use craft\base\Plugin as BasePlugin;
@@ -60,10 +61,12 @@ class Plugin extends BaseGenerator
     {
         $this->name = $this->command->prompt('Plugin name:', [
             'required' => true,
+            'validator' => fn(string $input) => StringHelper::isUtf8($input),
         ]);
 
         $this->developer = $this->command->prompt('Developer name:', [
             'required' => true,
+            'validator' => fn(string $input) => StringHelper::isUtf8($input),
         ]);
 
         $craftVersion = Craft::$app->getVersion();
@@ -111,7 +114,9 @@ EOD));
             'pattern' => sprintf('/%s/', self::PACKAGE_NAME_PATTERN),
         ]);
 
-        $this->description = $this->command->prompt('Plugin description:');
+        $this->description = $this->command->prompt('Plugin description:', [
+            'validator' => fn(string $input) => StringHelper::isUtf8($input),
+        ]);
 
         if (!$this->private) {
             $this->license = $this->command->select('How should the plugin be licensed?', [
@@ -120,12 +125,15 @@ EOD));
             ]);
 
             $this->email = $this->command->prompt('Support email:', [
-                'validator' => function(string $input, ?string &$error): bool {
-                    return (new EmailValidator())->validate($input, $error);
-                },
+                'validator' => fn(string $input, ?string & $error) =>
+                    (new EmailValidator())->validate($input, $error) &&
+                    StringHelper::isUtf8($input),
             ]);
 
-            $this->repo = $this->command->prompt('GitHub repo URL:');
+            $this->repo = $this->command->prompt('GitHub repo URL:', [
+                'validator' => fn(string $input) => StringHelper::isUtf8($input),
+            ]);
+
             if ($this->repo) {
                 $this->repo = StringHelper::toLowerCase($this->repo);
                 $this->repo = str_replace('http://', 'https://', $this->repo);
@@ -141,8 +149,17 @@ EOD));
             'pattern' => '/^[\d\.]+$/',
         ]);
 
+        if (
+            Comparator::greaterThanOrEqualTo($craftVersion, '4.4.0') &&
+            VersionParser::parseStability($craftVersion) === 'stable'
+        ) {
+            $defaultMinCraftVersion = preg_replace('/^(\d+\.\d+).*/', '$1.0', $craftVersion);
+        } else {
+            $defaultMinCraftVersion = $craftVersion;
+        }
+
         $this->minCraftVersion = $this->command->prompt('Minimum Craft CMS version:', [
-            'default' => $craftVersion,
+            'default' => $defaultMinCraftVersion,
             'validator' => function(string $input, ?string &$error) use ($minPrivatePluginVersion): bool {
                 if (!preg_match('/^[\d\.]+(-\w+(\.\d+)?)?$/', $input)) {
                     $error = 'Invalid version.';