+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 - Improved the validation error for private plugins’ handle prompts, for handles that don’t begin with an underscore. ([#18](https://github.com/craftcms/generator/issues/18))
+- Fixed a bug where it was possible to enter invalid UTF-8 characters into the plugin generator prompts, causing an exception when writing the `composer.json` file.
+- Craft CMS 4.4.11+ is now required.
 
 ## 1.4.0 - 2023-04-19
 - Generated plugins now include a “Create Release” GitHub action, which will create a new GitHub Release whenever the Craft Plugin Store is notified of a new version tag.