+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 - Improved the validation error for private plugins’ handle prompts, for handles that don’t begin with an underscore. ([#18](https://github.com/craftcms/generator/issues/18))
 - The plugin generator’s “Minimum Craft CMS version” prompt now defaults to the initial minor release of the current Craft version (e.g. `4.4.0` rather than `4.4.11`), if running a stable release of Craft 4.4 or later.
+- The plugin generator now warns against using non-unique plugin handles.
 - Generated plugins’ and modules’ `init()` methods now have a `void` return type. ([#20](https://github.com/craftcms/generator/issues/20)) 
 - Fixed a bug where it was possible to enter invalid UTF-8 characters into the plugin generator prompts, causing an exception when writing the `composer.json` file.
 - Craft CMS 4.4.11+ is now required.